Renamed src to lppdydsmc and restructured projects. Changed imports too so it is better. Tested it and it works.
