+++ lppydsmc/collision/__init__.py
@@ -1,0 +1,3 @@
+from . import collider
+
+from .collider import handler_particles_collisions
