+++ lppydsmc/data_structures/__init__.py
@@ -1,0 +1,3 @@
+from . import grid
+from .grid import Grid
+from .particle import Particle
