+++ lppydsmc/injection/__init__.py
@@ -1,0 +1,1 @@
+from .injector import inject
