+++ plotting/__init__.py
@@ -1,0 +1,3 @@
+from . import plt_tools # analysis uses plt_tools
+from . import analysis
+
