--- src/__init__.py
@@ -1,2 +1,0 @@
-from src import utils
-from src import data
