--- src/plotting/__init__.py
@@ -1,1 +1,0 @@
-from .matplotlib_utils import plot_boundaries, plot_particles, plot_grid, plot_system
