--- src/utils/__init__.py
@@ -1,8 +1,0 @@
-from .wall_collision import handler_wall_collision, make_collisions, make_collisions_vectorized, make_collisions_out_walls, handler_wall_collision_point, deal_with_corner, _reflect_particle
-from .physics import gaussian, maxwellian_mean_speed, maxwellian_flux, get_mass_part, mean_free_path, mean_free_time
-from .injector import inject
-from .schemes import euler_explicit, leap_frog
-from .advector import advect
-from .particle import Particle
-from .grid import Grid, pos_in_grid, convert_to_grid_datatype, convert_to_grid_format, default_hashing
-from .collider import candidates, index_choosen_couples, probability, is_colliding, reflect, handler_particles_collisions
